requirements.txt
scipy
pandas
plotly
networkx
pyyaml
torch
optuna
pytest
pytest-xdist
numba
psutil
py3nvml
pysr
lammps
ase
gymnasium
